backend/requirements.txt
# Finalyse — Backend Dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pdfplumber==0.11.0
Pillow==10.3.0
pydantic==2.7.1
python-dotenv==1.0.1

# Optional — PDF to image conversion (choose one):
# pypdfium2==4.30.0       # Recommended: no system dependencies
# pdf2image==1.17.0       # Alternative: requires Poppler

# Optional — better XLSX exports
# openpyxl==3.1.2
# reportlab==4.2.0

# Frontend Dependencies (install separately)
# PyQt6==6.7.0
